% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{collect_data}
\alias{collect_data}
\title{Collect all data from a CMS dataset}
\usage{
collect_data(data_id, select = NULL, ask_for_confirmation = TRUE)
}
\arguments{
\item{data_id}{dataset identifier, could be "Dataset Type Identifier" or "Dataset Version Identifier"}

\item{select}{a vector of column names to select, default is \code{NULL}}

\item{ask_for_confirmation}{whether to ask for confirmation before downloading, default is \code{TRUE}}
}
\value{
a tibble that contains all data from the dataset
}
\description{
Collect all data from a CMS dataset
}
\examples{
data_id <- '9767cb68-8ea9-4f0b-8179-9431abc89f11'
collect_data(data_id, ask_for_confirmation = FALSE)
aco_participants_all <- collect_data(data_id, ask_for_confirmation = FALSE)
aco_participants_id <- collect_data(data_id, select = c('aco_id'), ask_for_confirmation = FALSE)
}
