% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_datasets.R
\name{show_datasets}
\alias{show_datasets}
\title{Show all datasets from CMS}
\usage{
show_datasets()
}
\value{
A tibble for all latest public datasets from CMS. Columns include
\itemize{
\item \code{title}
\item \code{identifier}: the "Dataset Type Identifier", which doesn't change and always points to the latest version of the dataset. For "Dataset Version Identifier", visit its \code{landing_page}.
\item \code{temporal}: time period
\item \code{modified}: last modified date
\item \code{description}
\item \code{reference}: documentation link, named "describedBy" by CMS
\item \code{landing_page}: webpage for the dataset
}
}
\description{
Show the basic information of all \strong{latest} public datasets from Centers for Medicare & Medicaid Services (CMS) in the tibble format.
}
\examples{
cms_datasets <- show_datasets()
}
