% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_head_data.R
\name{slice_head_data}
\alias{slice_head_data}
\title{Slice the first n rows from a CMS dataset}
\usage{
slice_head_data(data_id, n = 10, select = NULL)
}
\arguments{
\item{data_id}{dataset identifier, could be "Dataset Type Identifier" or "Dataset Version Identifier"}

\item{n}{number of rows to slice, default is 100}

\item{select}{a vector of column names to select, default is \code{NULL}}
}
\value{
A tibble for the first n rows of the dataset
}
\description{
Slice the first n rows from a CMS dataset
}
\examples{
data_id <- '9767cb68-8ea9-4f0b-8179-9431abc89f11'
aco_participants <- slice_head_data(data_id)
aco_participants_100 <- slice_head_data(data_id, 100)
aco_participants_id <- slice_head_data(data_id, select = c('aco_id'))
}
